requirements.txt
fastapi
uvicorn[standard]

celery
redis

SQLAlchemy
psycopg2-binary

python-dotenv
pydantic

